Change load and reload
The <input> architecture did not allow me to easily do a reload as a
file changed. In order to see the changes in a file, I'd have to load
a second model and then load the original model again.

By clearing the input value, I could reduce that by one step; simply select
the file again. It would reload. But, I lost the display of the file name.

BY doing this, I have the file name *and* the one-click reload.